backend/requirements.txt
asgiref>=3.6.0
Django==3.2.24
django-ranged-response==0.2.0
PyMysql==1.1.0
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
werkzeug==3.0.0
objtyping==0.5.4
pyjwt==2.8.0
openpyxl==3.0.9
pandas>=1.3.0
numpy>=1.21.0
tensorflow>=2.6.0
joblib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
djangorestframework==3.14.0
jieba>=0.42.1
scikit-learn>=0.24.2
networkx>=2.6.0
neo4j>=4.4.0
scipy>=1.7.0
torch>=1.9.0
transformers>=4.10.0
python-dateutil>=2.8.2
matplotlib>=3.4.0
seaborn>=0.11.0